+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 using `'param_groups'` key in model components.
 - support for None `additional_input` shape in model for scalar tensor input.
 - support for image size type of list (w,h), or non-square image, in predict and export.
+- validation loss calculation to experiment logger
+
 
 ### Changed
 - model checkpoint not save on hyperparameter optimization